+++ MASTER_S3-P4/S3/iMakie-ESP32_S3_EXTENDER/src/config.h
@@ -12,7 +12,7 @@
 #elif defined(DEVICE_S3_EXTENDER)
     #define DEVICE_FAMILY       0x14
     #define VERSION_REPLY_CMD   0x14
-    #define NUM_SLAVES          4   // TESTING=5 | PRODUCCIÓN=8 — no cambiar aquí sin hardware real
+    #define NUM_SLAVES          1   // TESTING=5 | PRODUCCIÓN=8 — no cambiar aquí sin hardware real
 
 #else
     #error "DEBE DEFINIR: DEVICE_P4_MASTER o DEVICE_S3_EXTENDER en platformio.ini build_flags"